# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Gaze Engine API
  version: 3.0.0
  description: Privacy-first local video search engine API
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "48100"
security:
  - bearerAuth: []
paths:
...
  /libraries:
    get:
      operationId: listLibraries
      summary: List all libraries
      responses:
        "200":
          description: List of libraries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibrariesResponse"
    post:
      operationId: addLibrary
      summary: Add a library (folder path)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLibraryRequest"
      responses:
        "200":
          description: Library added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
  /libraries/{id}:
    get:
      operationId: getLibrary
      summary: Get library details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Library details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
    delete:
      operationId: deleteLibrary
      summary: Remove library and all indexed data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Library deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /libraries/{id}/scan:
    post:
      operationId: scanLibrary
      summary: Trigger rescan for new/changed/deleted files
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
...
  /assets/thumbnail:
    get:
      operationId: getThumbnail
      summary: Serve a thumbnail image from the thumbnails directory
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thumbnail image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /jobs/{id}:
    get:
      operationId: getJob
      summary: Get job status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: cancelJob
      summary: Cancel job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /jobs/start:
    post:
      operationId: startIndexing
      summary: Start indexing for queued videos
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Indexing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
...
  /assets/face:
    get:
      operationId: getFaceCrop
      summary: Serve a face crop image
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Absolute face crop path
      responses:
        "200":
          description: Face crop image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "403":
          description: Forbidden
        "404":
          description: Face crop not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Library:
      type: object
      required:
        - library_id
        - folder_path
        - created_at_ms
      properties:
        library_id:
          type: string
        folder_path:
          type: string
        name:
          type: string
        recursive:
          type: boolean
        video_count:
          type: integer
        indexed_count:
          type: integer
        created_at_ms:
          type: integer
    LibrariesResponse:
      type: object
      required:
        - libraries
      properties:
        libraries:
          type: array
          items:
            $ref: "#/components/schemas/Library"
    AddLibraryRequest:
      type: object
      required:
        - folder_path
      properties:
        folder_path:
          type: string
        name:
          type: string
        recursive:
          type: boolean
          default: true
    ScanResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [started, already_scanning]
...
    Job:
      type: object
      required:
        - job_id
        - video_id
        - status
        - created_at_ms
      properties:
        job_id:
          type: string
        video_id:
          type: string
        status:
          type: string
          enum: [PENDING, EXTRACTING_AUDIO, TRANSCRIBING, EXTRACTING_FRAMES, EMBEDDING, DETECTING, DONE, FAILED, CANCELLED]
        current_stage:
          type: string
        progress:
          type: number
          minimum: 0
          maximum: 1
        message:
          type: string
        error_code:
          type: string
        error_message:
          type: string
        created_at_ms:
          type: integer
        updated_at_ms:
          type: integer
...